refactor(app): :recycle: separates index file into api.js and server.js
api.js: starts the app with express and uses express.json() to read JSON

server.js: uses cors, error middleware and the routes and listen on port 3000
