+++ api.js
@@ -1,0 +1,7 @@
+const express = require('express');
+
+const app = express();
+
+app.use(express.json());
+
+module.exports = app;
